# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traveloop Travel Planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips for current user
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{id}/budget:
    get:
      operationId: getTripBudget
      tags: [trips]
      summary: Get budget breakdown for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetBreakdown"
  /trips/{id}/public:
    get:
      operationId: getPublicTrip
      tags: [trips]
      summary: Get public view of trip (no auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public trip view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
  /trips/{id}/share:
    patch:
      operationId: updateTripShare
      tags: [trips]
      summary: Toggle public sharing for trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/copy:
    post:
      operationId: copyTrip
      tags: [trips]
      summary: Copy a public trip to own account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Copied trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"

  # STOPS
...
components:
  schemas:
...
    Trip:
      type: object
      required: [id, name, startDate, endDate, isPublic, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverPhoto:
          type: ["string", "null"]
        isPublic:
          type: boolean
        stopCount:
          type: integer
        totalBudget:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripDetail:
      type: object
      required: [id, name, startDate, endDate, isPublic, stops]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverPhoto:
          type: ["string", "null"]
        isPublic:
          type: boolean
        totalBudget:
          type: ["number", "null"]
        stops:
          type: array
          items:
            $ref: "#/components/schemas/StopDetail"
        createdAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverPhoto:
          type: string
        totalBudget:
          type: number
    TripUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverPhoto:
          type: ["string", "null"]
        totalBudget:
          type: ["number", "null"]
    ShareUpdate:
      type: object
      required: [isPublic]
      properties:
        isPublic:
          type: boolean
...
    BudgetBreakdown:
      type: object
      required: [totalBudget, totalEstimated, breakdown, dailyCosts, overBudgetDays]
      properties:
        totalBudget:
          type: ["number", "null"]
        totalEstimated:
          type: number
        breakdown:
          type: object
          properties:
            transport:
              type: number
            accommodation:
              type: number
            activities:
              type: number
            meals:
              type: number
            other:
              type: number
        dailyCosts:
          type: array
          items:
            type: object
            required: [date, cost]
            properties:
              date:
                type: string
                format: date
              cost:
                type: number
        overBudgetDays:
          type: array
          items:
            type: string
            format: date

    # STOP SCHEMAS
...
    StopDetail:
      type: object
      required: [id, tripId, cityId, arrivalDate, departureDate, orderIndex, activities]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        city:
          $ref: "#/components/schemas/City"
        arrivalDate:
          type: string
          format: date
        departureDate:
          type: string
          format: date
        orderIndex:
          type: integer
        accommodation:
          type: ["string", "null"]
        accommodationCost:
          type: ["number", "null"]
        transportCost:
          type: ["number", "null"]
        mealsCost:
          type: ["number", "null"]
        activities:
          type: array
          items:
            $ref: "#/components/schemas/StopActivity"
...
    StopActivity:
      type: object
      required: [id, stopId, activityId, activity]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityId:
          type: integer
        scheduledTime:
          type: ["string", "null"]
        activity:
          $ref: "#/components/schemas/Activity"
...
    City:
      type: object
      required: [id, name, country, region]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        region:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        costIndex:
          type: ["number", "null"]
        popularityScore:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]

    # ACTIVITY SCHEMAS
    Activity:
      type: object
      required: [id, cityId, name, type, estimatedCost, durationHours]
      properties:
        id:
          type: integer
        cityId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [sightseeing, food, adventure, culture, shopping, wellness, nightlife, other]
        estimatedCost:
          type: number
        durationHours:
          type: number
        imageUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]

    # CHECKLIST SCHEMAS
...