# lib/api-zod/src/generated/types/budgetBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop Travel Planning API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetBreakdownBreakdown } from "./budgetBreakdownBreakdown";
import type { BudgetBreakdownDailyCostsItem } from "./budgetBreakdownDailyCostsItem";

export interface BudgetBreakdown {
  /** @nullable */
  totalBudget: number | null;
  totalEstimated: number;
  breakdown: BudgetBreakdownBreakdown;
  dailyCosts: BudgetBreakdownDailyCostsItem[];
  overBudgetDays: Date[];
}