# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop Travel Planning API
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  startDate?: Date;
  endDate?: Date;
  /** @nullable */
  coverPhoto?: string | null;
  /** @nullable */
  totalBudget?: number | null;
}